requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.3
drf-yasg>=1.21
python-dotenv>=1.0
Pillow>=10.0
psycopg2-binary>=2.9
